BDD/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
kafka-python